1. add test cases for calculate_bboxes_area, denormalize_bboxes_np and normalize_bboxes_np. 2. add test cases for test_bboxes_flip, test_crop_bboxes_by_coords, test_bboxes_center_crop, test_bboxes_crop, test_bboxes_random_crop. 3. Add methods in benchmark_bbox.py.

+++ albumentations/augmentations/crops/functional.py
@@ -123,8 +123,8 @@ def bboxes_random_crop(
     bboxes: np.ndarray,
     crop_height: int,
     crop_width: int,
-    h_start: int,
-    w_start: int,
+    h_start: float,
+    w_start: float,
     rows: int,
     cols: int,
 ) -> np.ndarray:
